+++ r/NEWS.md
@@ -19,8 +19,10 @@
 
 # arrow 0.16.0.9000
 
+
 * `write_feather`, `write_arrow` and `write_parquet` now return their input
 similar to `write_*` functions from `readr` (#7796 @boshek)
+* Dataset filtering is now correctly supported for all Arrow date/time/timestamp column types.
 
 # arrow 0.16.0